Added COBie to IFC Transformation Settings.  (Safety Commit)
This commit includes a new COBie to IFC settings file that allows you to
specify the list of ignored entity types, attribute names to ignore,
etc.  This commit is being made for safety.  Functionality is still
being tested and a release is forthcoming.

+++ COBieShared/src/org/bimserver/cobie/shared/utility/PluginRuntimeFileHelper.java
@@ -22,9 +22,11 @@ public class PluginRuntimeFileHelper
 	private static final Logger LOGGER = LoggerFactory
 			.getLogger(PluginRuntimeFileHelper.class);
 
+	public enum Persistence { PERMANENT, TEMP}
+	
 
-    public static File prepareSerializerConfigFile(PluginManager pluginManager,
-            String name, Plugin plugin, String configPath) throws IOException
+    public static File prepareSerializerResource(PluginManager pluginManager,
+            String name, Plugin plugin, String configPath, Persistence persistence) throws IOException
     {
         PluginContext pluginContext = pluginManager.getPluginContext(plugin);
         Path config = pluginContext.getRootPath().resolve(configPath);
@@ -35,19 +37,27 @@ public class PluginRuntimeFileHelper
         {
             // File tmpFolder = new
             // File(pluginManager.getTempDir(),TEMP_DIRECTORY_NAME);
-        	
-            File tmpFolder = new File(pluginManager.getTempDir().toString());
-            File tempChildFolder = new File(tmpFolder, name);
+        	File folder = getParentFolder(pluginManager, plugin, persistence);
+            File tempChildFolder = new File(folder, name);
             try
             {
-                if (tempChildFolder.exists())
+                if (tempChildFolder.exists() && persistence == Persistence.TEMP)
                 {
                     FileUtils.forceDelete(tempChildFolder);
+                    
                 }
-                FileUtils.forceMkdir(tempChildFolder);
+                if(!tempChildFolder.exists())
+                {
+                	FileUtils.forceMkdir(tempChildFolder);
+                }
+                
                 configurationFile = new File(tempChildFolder, tempFileName);
-                IOUtils.copy(inputStream, new FileOutputStream(
-                        configurationFile));
+                if(!configurationFile.exists() || persistence == Persistence.TEMP)
+                {
+                    IOUtils.copy(inputStream, new FileOutputStream(
+                            configurationFile));
+                }
+
             }
             catch (IOException e)
             {
@@ -57,23 +67,46 @@ public class PluginRuntimeFileHelper
         return configurationFile;
     }
 
-    public static HashMap<String, File> prepareSerializerConfigFiles(
+	private static File getParentFolder(PluginManager pluginManager,
+			Plugin plugin, Persistence persistence) 
+	{
+		File folder;
+		switch(persistence)
+		{
+			case PERMANENT:
+				folder = new File(pluginManager.getTempDir().getParent().toString());
+				break;
+			case TEMP:
+				folder = new File(pluginManager.getTempDir().toString());
+				break;
+			default:
+				folder = new File(pluginManager.getTempDir().toString());
+				break;
+		}
+		return folder;
+	}
+
+    public static HashMap<String, File> prepareSerializerResource(
             PluginManager pluginManager, String name, Plugin plugin,
-            ArrayList<String> configPaths) throws IOException
+            ArrayList<String> configPaths, Persistence persistence) throws IOException
     {
         PluginContext pluginContext = pluginManager.getPluginContext(plugin);
         // File tmpFolder = new File(pluginManager.getTempDir(),
         // TEMP_DIRECTORY_NAME);
-        File tmpFolder = new File(pluginManager.getTempDir().toString());
-        File tempChildFolder = new File(tmpFolder, name);
+        File folder = getParentFolder(pluginManager, plugin, persistence);
+        File tempChildFolder = new File(folder, name);
         HashMap<String, File> configFiles = new HashMap<String, File>();
         try
         {
-            if (tempChildFolder.exists())
+            if (tempChildFolder.exists() && persistence == Persistence.TEMP)
             {
                 FileUtils.forceDelete(tempChildFolder);
             }
-            FileUtils.forceMkdir(tempChildFolder);
+            if (!tempChildFolder.exists())
+            {
+            	FileUtils.forceMkdir(tempChildFolder);
+            }
+            
         }
         catch (IOException e)
         {
@@ -93,9 +126,11 @@ public class PluginRuntimeFileHelper
             {
                 if (inputStream != null)
                 {
-
-                    IOUtils.copy(inputStream, new FileOutputStream(
-                            configurationFile));
+                	if(!configurationFile.exists() || persistence == Persistence.TEMP)
+                	{
+                        IOUtils.copy(inputStream, new FileOutputStream(
+                                configurationFile));
+                	}
                     configFiles.put(configPath, configurationFile);
                 }
                 else
